refactor(voice)!: trim #707 voice slop to minimum-viable (#715)
Group A subset of the voice #707 follow-up cleanup. Removes dead surface,
dedupes repeated comments, collapses unreachable test scaffolding, extracts
duplicated test harness into shared fixtures, fixes self-skip discipline,
and strengthens weak assertions. No runtime behavior change.

- Delete the hollow source-reading guard test and the JSDoc-text
  assertions, keeping the behavioral coverage [items 1, 2].
- Collapse the unreachable try/catch/null theater around scenario.run()
  in the kitchen-sink test [item 5].
- State each repeated comment once: the EL idle-prompt story, the realtime
  instructions-override note, and the USER_TURN_NO_AUDIO essay [item 6].
- Delete dead surface: the deprecated no-op silenceTailBytes option, the
  unused VOICE_COHERENCE_CRITERIA test constant, and the vestigial EL url
  getter + ELEVENLABS_CONVAI_URL_TEMPLATE. Zero references verified before
  removal; sendText was found still referenced and kept [item 7].
- Extract FakeWebSocket/makeFakeConv, the mock realtime ws-server, and the
  realtimeUser() helper + shared instructions into __tests__/fixtures and
  helpers/ [item 8].
- Convert if(!hasKey) return self-skips to it.skipIf so skipped hosted
  tests report SKIPPED, not a hollow PASSED [item 9].
- Import the shared AGENTS_HEARD_EACH_OTHER judge constant instead of
  bespoke criteria in two e2e files [item 10].
- Rename the repro-705 scratch file to an intention-revealing name, trim
  the kitchen-sink war-story header, and strengthen the multi-turn
  assertions to turn counts plus manifest consistency [item 12].

Out of scope, split per the issue: dynamicVariables/deepMerge (item 3, own
PR), the autonomous realtime-user adapter (item 4, ships with #711), e2e
consolidation (item 11), and groups B/C/F.

Refs #715

BREAKING CHANGE: the ElevenLabs adapter's vestigial `url` getter and the
`ELEVENLABS_CONVAI_URL_TEMPLATE` export are removed, along with the
deprecated no-op `silenceTailBytes` option on ElevenLabsAgentAdapterOptions.
None were load-bearing (the SDK dials its own signed socket and the
continuous mic pump replaced the bounded silence tail), but they were part
of the public voice surface.
