+++ javascript/src/execution/scenario-execution.ts
@@ -1262,14 +1262,9 @@ export class ScenarioExecution implements ScenarioExecutionLike, VoiceExecutorSt
    * @param role - The role the agent was called for.
    * @returns The messages to add/broadcast — voiced where applicable.
    * @throws {Error} ({@link USER_TURN_NO_AUDIO_FOR_VOICE_AUT}, the fail-closed
-   *   invariant) when the FINAL (post-voiceify) user turn for a voice agent
-   *   under test carries no audio. Adapter-AGNOSTIC and strictly stronger than
-   *   the old realtime-user type-check it replaced: it trips on the produced
-   *   ARTIFACT (a no-audio turn) regardless of producer type — catching a
-   *   realtime adapter that returns text AND a non-realtime/non-voice-sim
-   *   producer the type-check let through silently — rather than degrade the
-   *   user side to a text turn the voice agent can't hear. The autonomous
-   *   OpenAI Realtime user PASSES it (its `call()` returns audio).
+   *   invariant) when the FINAL (post-voiceify) user turn for a voice agent under
+   *   test carries no content. See {@link USER_TURN_NO_AUDIO_FOR_VOICE_AUT} for the
+   *   full rationale.
    */
   private async voiceifyGeneratedUserTurn(
     messages: ModelMessage[],
@@ -1308,20 +1303,16 @@ export class ScenarioExecution implements ScenarioExecutionLike, VoiceExecutorSt
       outgoing = voiced;
     }
 
-    // Fail-closed invariant (adapter-AGNOSTIC): a USER turn for a voice agent
-    // under test MUST carry REAL content — audio bytes the AUT can hear, or a
-    // transcript (for a text-commit adapter) — the why is on
-    // USER_TURN_NO_AUDIO_FOR_VOICE_AUT.
+    // Fail-closed invariant: a USER turn for a voice agent under test MUST carry
+    // REAL content (audio bytes the AUT can hear, or a transcript for a text-commit
+    // adapter); the why is on USER_TURN_NO_AUDIO_FOR_VOICE_AUT.
     // LOAD-BEARING ORDER: this runs AFTER the voiceify/TTS step, on the FINAL
-    // outgoing turn — so a legitimate voice-user-sim TEXT turn (text is PRE-TTS)
-    // is allowed through ONCE voiced, while a realtime adapter that returns text,
-    // or any producer that yields a CONTENT-LESS user turn, fails loud rather
-    // than silently degrading the agent under test to a turn it can't hear.
-    // NB: test real content, not `messageHasAudio` — a ZERO-byte audio part still
+    // outgoing turn, so a legitimate voice-user-sim TEXT turn (text is PRE-TTS) is
+    // allowed through ONCE voiced, while a realtime adapter that returns text, or
+    // any producer that yields a CONTENT-LESS user turn, fails loud.
+    // NB: test real content, not `messageHasAudio`: a ZERO-byte audio part still
     // "has audio" (extractAudio is non-null), so an empty-audio #708-flake turn
     // would sail through and feed the AUT silence (a receiveAudio-timeout hang).
-    // It is the produced ARTIFACT, not the producer's TYPE, that trips it —
-    // strictly stronger than the old isRealtimeUserAgent check.
     const carriesContent = outgoing.some((message) => {
       const audio = extractAudio(message);
       return (
