+++ javascript/src/voice/index.ts
@@ -207,7 +207,6 @@ export {
   ComposableVoiceAgent,
   ElevenLabsAgentAdapter,
   ElevenLabsVoiceAgent,
-  ELEVENLABS_CONVAI_URL_TEMPLATE,
   type ComposableVoiceAgentOptions,
   type ElevenLabsAgentAdapterOptions,
   type ElevenLabsVoiceAgentOptions,
